# schemas/openapi/edge-agent-local-api.yaml
openapi: 3.1.0
info:
  title: FCC Middleware — Edge Agent Local API
  version: 1.0.0
  description: |
    Local REST API served by the Edge Agent (Ktor embedded server) on the HHT device.
    Consumed by Odoo POS running on the same device (localhost) and optionally
    by secondary HHTs on the same WiFi LAN.

    ## Conventions

    - **Dates:** UTC ISO 8601 everywhere (`2026-03-11T14:00:00Z`). Timezone only in the display layer.
    - **Currency:** Integer minor units (e.g., cents). ISO 4217 `currencyCode` alongside.
    - **Volume:** Microlitres (integer). 1 L = 1 000 000 µL.
    - **IDs:** UUID v4 for middleware-generated IDs. FCC-vendor IDs preserved as opaque strings.
    - **JSON field names:** camelCase.
    - **Pagination:** Cursor-based (`cursor` + `pageSize`). Default 50, max 200.
    - **Error envelope:** `{ errorCode, message, details, traceId, timestamp }`.

    ## Authentication

    - **Localhost (127.0.0.1):** No API key required. All requests from the loopback address
      are trusted. This covers Odoo POS running on the same HHT.
    - **LAN (other HHTs):** `X-Api-Key` header with the site LAN API key. Required when
      `localApi.enableLanApi = true` in `SiteConfig`. The LAN API key is provisioned
      via `localApi.lanApiKeyRef`.

  contact:
    name: Platform Team
servers:
  - url: http://127.0.0.1:{port}
    description: Localhost — Odoo POS on the same HHT (no auth required)
    variables:
      port:
        default: "8080"
...
    variables:
...
paths:
...
  /api/v1/transactions:
    get:
      operationId: listLocalTransactions
      tags: [Transactions]
      summary: List buffered transactions
      description: |
        Returns locally buffered canonical transactions with optional filtering.
        Useful for Odoo POS to reconcile local state or display recent activity.

        Results are ordered by `completedAt` descending (most recent first).
      security:
        - LanApiKey: []
        - {}
      parameters:
        - $ref: '#/components/parameters/CursorParam'
        - $ref: '#/components/parameters/PageSizeParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, SYNCED, SYNCED_TO_ODOO, STALE_PENDING, DUPLICATE, ARCHIVED]
          description: Filter by local sync status.
        - name: pumpNumber
          in: query
          schema:
            type: integer
            minimum: 1
          description: Filter to a specific pump.
        - name: from
          in: query
          schema:
            type: string
            format: date-time
          description: "Filter: `completedAt >= from` (UTC ISO 8601)."
        - name: to
          in: query
          schema:
            type: string
            format: date-time
          description: "Filter: `completedAt <= to` (UTC ISO 8601)."
      responses:
        "200":
          description: Paginated list of buffered transactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocalTransactionPage'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /api/v1/transactions/{id}:
    get:
      operationId: getLocalTransactionById
      tags: [Transactions]
      summary: Get transaction by ID
      description: Returns a single locally buffered transaction by its middleware UUID.
      security:
        - LanApiKey: []
        - {}
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Middleware UUID of the transaction.
      responses:
        "200":
          description: Transaction found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocalTransaction'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /api/v1/transactions/{id}/acknowledge:
    post:
      operationId: acknowledgeLocalTransaction
      tags: [Transactions]
      summary: Acknowledge a transaction — stamp odooOrderId
      description: |
        Called by Odoo POS to stamp a locally buffered transaction with its Odoo order
        reference. This does not immediately change cloud sync status; the `odooOrderId`
        will be included in the next cloud upload batch or reported on the next cloud
        status poll.

        Idempotent: re-acknowledging with the same `odooOrderId` returns the transaction
        unchanged. Acknowledging with a different `odooOrderId` returns 409 Conflict.
      security:
        - LanApiKey: []
        - {}
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [odooOrderId]
              additionalProperties: false
              properties:
                odooOrderId:
                  type: string
                  minLength: 1
                  maxLength: 100
                  description: Odoo POS order reference to stamp on the transaction.
      responses:
        "200":
          description: Transaction acknowledged (or already acknowledged with same odooOrderId)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LocalTransaction'
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          $ref: '#/components/responses/NotFound'
        "409":
          description: Transaction already acknowledged with a different odooOrderId
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "500":
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    LanApiKey:
      type: apiKey
      in: header
      name: X-Api-Key
      description: |
        LAN API key required for requests from non-primary HHTs on the same WiFi network.
        Not required for localhost (127.0.0.1) requests from Odoo POS on the same device.
        The key is provisioned via `localApi.lanApiKeyRef` in SiteConfig and stored in
        Android Keystore on the primary HHT.
  parameters:
    CursorParam:
      name: cursor
      in: query
      schema:
        type: string
      description: Opaque cursor from a previous response's `nextCursor`. Omit for first page.
    PageSizeParam:
      name: pageSize
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
      description: Number of records per page. Max 200 for the local API.
  responses:
    BadRequest:
      description: Invalid request body or query parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            errorCode: VALIDATION_ERROR
            message: "requestedAmount must be >= 1"
            details:
              field: requestedAmount
              value: 0
            traceId: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
            timestamp: "2026-03-11T14:00:00Z"
    Unauthorized:
      description: Missing or invalid LAN API key (LAN access only; not applicable on localhost)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found in local buffer
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Unexpected agent error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      description: Standard error envelope. Consistent with Cloud API error format.
      required: [errorCode, message, traceId, timestamp]
      additionalProperties: false
      properties:
        errorCode:
          type: string
          description: |
            Machine-readable error code. Values:
            - `VALIDATION_ERROR` — invalid or missing fields
            - `NOT_FOUND` — resource not in local buffer
            - `CONFLICT` — state conflict (duplicate pre-auth, wrong odooOrderId, terminal state)
            - `UNAUTHORIZED` — invalid LAN API key
            - `FCC_UNREACHABLE` — FCC not responding
            - `FCC_TIMEOUT` — FCC response timeout
            - `INTERNAL_ERROR` — unexpected agent error
        message:
          type: string
        details:
          type: object
          additionalProperties: true
          nullable: true
        traceId:
          type: string
          format: uuid
        timestamp:
          type: string
          format: date-time

    # ── Pagination ──────────────────────────────────────────────────────────
    PageMeta:
      type: object
      required: [pageSize, hasMore]
      additionalProperties: false
      properties:
        pageSize:
          type: integer
        hasMore:
          type: boolean
        nextCursor:
          type: string
          nullable: true

    # ── Local transaction ───────────────────────────────────────────────────
    LocalTransaction:
      type: object
      description: |
        Locally buffered canonical transaction (subset of cloud CanonicalTransaction).
        `rawPayloadJson` and cloud-only fields are excluded.
      required:
        - id
        - fccTransactionId
        - siteCode
        - pumpNumber
        - nozzleNumber
        - productCode
        - volumeMicrolitres
        - amountMinorUnits
        - unitPriceMinorPerLitre
        - startedAt
        - completedAt
        - fccVendor
        - currencyCode
        - status
        - ingestedAt
        - updatedAt
        - isDuplicate
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        fccTransactionId:
          type: string
          minLength: 1
          maxLength: 256
        siteCode:
          type: string
        pumpNumber:
          type: integer
          minimum: 1
        nozzleNumber:
          type: integer
          minimum: 1
        productCode:
          type: string
        volumeMicrolitres:
          type: integer
          exclusiveMinimum: 0
        amountMinorUnits:
          type: integer
          exclusiveMinimum: 0
        unitPriceMinorPerLitre:
          type: integer
          exclusiveMinimum: 0
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        fccVendor:
          type: string
          enum: [DOMS, RADIX, ADVATEC, PETRONITE]
        currencyCode:
          type: string
          pattern: '^[A-Z]{3}$'
        status:
          type: string
          enum: [PENDING, SYNCED, SYNCED_TO_ODOO, STALE_PENDING, DUPLICATE, ARCHIVED]
        ingestedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        isDuplicate:
          type: boolean
        odooOrderId:
          type: string
          nullable: true
          maxLength: 100
        attendantId:
          type: string
          nullable: true
          maxLength: 100
        fiscalReceiptNumber:
          type: string
          nullable: true
          maxLength: 100
    LocalTransactionPage:
      type: object
      required: [data, meta]
      additionalProperties: false
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/LocalTransaction'
        meta:
          $ref: '#/components/schemas/PageMeta'

    # ── Pump status ─────────────────────────────────────────────────────────
...